Move PromQL API error translation to querier registration.
