+++ pkg/api/api.go
@@ -269,7 +269,7 @@ func (a *API) RegisterQuerier(
 ) http.Handler {
 	api := v1.NewAPI(
 		engine,
-		queryable,
+		errorTranslateQueryable{queryable}, // Translate errors to errors expected by API.
 		func(context.Context) v1.TargetRetriever { return &querier.DummyTargetRetriever{} },
 		func(context.Context) v1.AlertmanagerRetriever { return &querier.DummyAlertmanagerRetriever{} },
 		func() config.Config { return config.Config{} },
